Enabled pool, client app, and identity roles
For user-pool and user-pool-client id's, changed grep to greedliy capture zero to all characters after the key-value pair. Returned JSON sometimes had ID values as the last pair, hence no `,` which would return a NULL value.

The `set-identity-pool-roles` requires an AWS account number as part of the ARN. Added `AWS_ACCOUNT` variable, which should be set to the account this script is run against.