Changed grep regex to OS X supported
OS X doesn't support perl style regex. Assumption is that for each returned JSON, there is only one "Id": or "ClientId":
line per file of interest. Reduced the regex to only look for that. Works in OS X.

Wonder if this file can be replaced with a CloudFormation template to do this, maybe with a Lambda helper, if needed.